+++ src/qibocal/protocols/characterization/allxy/allxy.py
@@ -51,14 +51,6 @@ class AllXYData(Data):
         else:
             self.data[qubit] = np.rec.array(ar)
 
-    @property
-    def qubits(self):
-        """Access qubits from data structure."""
-        return [q for q in self.data]
-
-    def __getitem__(self, qubit):
-        return self.data[qubit]
-
     def save(self, path):
         """Store results."""
         self.to_npz(path, self.data)
