+++ src/qibocal/protocols/characterization/flux_depedence/resonator_flux_dependence.py
@@ -73,14 +73,6 @@ class ResonatorFluxData(Data):
         ar["phase"] = phase
         self.data[qubit] = np.rec.array(ar)
 
-    @property
-    def qubits(self):
-        """Access qubits from data structure."""
-        return [q for q in self.data]
-
-    def __getitem__(self, qubit):
-        return self.data[qubit]
-
     def save(self, path):
         """Store results."""
         self.to_npz(path, self.data)
