+++ src/qibocal/protocols/characterization/resonator_punchout.py
@@ -1,4 +1,4 @@
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass, field
 from statistics import mode
 from typing import Dict, Optional
 
@@ -89,25 +89,6 @@ class ResonatorPunchoutData(Data):
         ar["msr"] = msr
         ar["phase"] = phase
         self.data[qubit] = np.rec.array(ar)
-
-    @property
-    def qubits(self):
-        """Access qubits from data structure."""
-        return [q for q in self.data]
-
-    @property
-    def global_params_dict(self):
-        global_dict = asdict(self)
-        global_dict.pop("data")
-        return global_dict
-
-    def __getitem__(self, qubit):
-        return self.data[qubit]
-
-    def save(self, path):
-        """Store results."""
-        self.to_json(path, self.global_params_dict)
-        self.to_npz(path, self.data)
 
 
 def _acquisition(
